requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
gspread==5.12.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.103.0
requests==2.31.0
python-dotenv==1.0.0 